More Ammo
Finished the first half of the Ammo display code, being the first 8
shots that are displayed based off of the LEDs at the top of the
device. Before there wasn't any other code to it and I also accidently
coded all of the different posibilities with = instead of == so it ran
through all of them instantaniously. I'm glad I caught such a mistake
quickly and was able to remedy it. The other half of the ammo display
code needs to be made, being the code for the last 8 shots displayed on
the screen .