Cursor, Ammo, Firing
I switched my code for the cursor to be effecient based on the example
we were shown in class. Previously the cursor was drawn and displayed
and on the loops when it blinked away I would clear the slate and draw
everything else again. I also edited my Ammo code. I previously had 8
copy pasted for statements when I really only needed one applying to
all of them which is what I switched it too. The game also now responds
to pressing the A button to fire. Nothing happens on screen, but the
LEDs register and delete one by one when a shot is fired. The PxAmmo
still needs to be fixed to display firing.