Boat, Ammo
I decided to approach the problem of randomly making boats in a
different way. To make it easier I decided to try and simply place
precreated boats randomly instead. I'm still having problems checking
and resetting the boats if they overlap and such. I was also having
trouble with displaying the last set of ammo. I tried to use the serial
display to test bugs but it also seemed to create some of its own. The
top row would only blinks when the serial is on so I'm not sure what
the problem is.