Broken Cursor Drawing
We tried all class today to make the cursor draw blue dots when a shot
is fired. We tried removing the clear slate and drawing the cursor dark
every other time just to test it. It worked and blinked however the
cursor now only draws a white dot everytime as oppose to the color we
wanted it to draw.