Check for hit, boolean problems
I now have good code to check if the cursor is over a part of the ship.
However, like my out of shots boolean variable, through testing using
the Serial monitor, we've discovered that the boolean isn't properly
returning its value. I'm working on converting the boolean to an int to
return either 1 or 0 instead.